+++ src/WallpaperGenerator/album_image.py
@@ -1,0 +1,1 @@
+from PIL import Image, ImageDraw
