+++ src/WallpaperGenerator/gradient.py
@@ -1,0 +1,2 @@
+from PIL import Image, ImageDraw
+
